+++ Thesis/Controllers/DishesController.cs
@@ -14,67 +14,90 @@ namespace Thesis.Controllers
     {
         private Entities db = new Entities();
 
+        /// <summary>
+        /// Az aktuális bejelentkezett user neve.
+        /// </summary>
+        private string currentUserName { get { return HttpContext.User.Identity.Name; } }
+
         // GET: Dishes
         public ActionResult Index()
         {
-            var dishes = db.Dishes.Include(d => d.DishCategory);
-            return View(dishes.ToList());
+            if (AuthenticationTools.UserHasRole(currentUserName, "StaffMember,Boss,Admin"))
+            {
+                var dishes = db.Dishes.Include(d => d.DishCategory);
+                return View(dishes.ToList());
+            }
+            return RedirectToAction("Login", "Account");            
         }
 
         // GET: Dishes/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (AuthenticationTools.UserHasRole(currentUserName, "StaffMember,Boss,Admin"))
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Dish dish = db.Dishes.Find(id);
+                if (dish == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(dish);
             }
-            Dish dish = db.Dishes.Find(id);
-            if (dish == null)
-            {
-                return HttpNotFound();
-            }
-            return View(dish);
+            return RedirectToAction("Login", "Account");
         }
 
         // GET: Dishes/Create
         public ActionResult Create()
         {
-            ViewBag.DishCategoryID = new SelectList(db.DishCategories, "ID", "Name");
-            return View();
+            if (AuthenticationTools.UserHasRole(currentUserName, "StaffMember,Boss,Admin"))
+            {
+                ViewBag.DishCategoryID = new SelectList(db.DishCategories, "ID", "Name");
+                return View();
+            }
+            return RedirectToAction("Login", "Account");
         }
 
         // POST: Dishes/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "ID,Name,Price,DishCategoryID")] Dish dish)
+        public ActionResult Create([Bind(Include = "Name,Price,DishCategoryID")] Dish dish)
         {
-            if (ModelState.IsValid)
+            if (AuthenticationTools.UserHasRole(currentUserName, "StaffMember,Boss,Admin"))
             {
-                db.Dishes.Add(dish);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
+                if (ModelState.IsValid)
+                {
+                    db.Dishes.Add(dish);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
 
-            ViewBag.DishCategoryID = new SelectList(db.DishCategories, "ID", "Name", dish.DishCategoryID);
-            return View(dish);
+                ViewBag.DishCategoryID = new SelectList(db.DishCategories, "ID", "Name", dish.DishCategoryID);
+                return View(dish);
+            }
+            return RedirectToAction("Login", "Account");
         }
 
         // GET: Dishes/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (AuthenticationTools.UserHasRole(currentUserName, "StaffMember,Boss,Admin"))
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Dish dish = db.Dishes.Find(id);
+                if (dish == null)
+                {
+                    return HttpNotFound();
+                }
+                ViewBag.DishCategoryID = new SelectList(db.DishCategories, "ID", "Name", dish.DishCategoryID);
+                return View(dish);
             }
-            Dish dish = db.Dishes.Find(id);
-            if (dish == null)
-            {
-                return HttpNotFound();
-            }
-            ViewBag.DishCategoryID = new SelectList(db.DishCategories, "ID", "Name", dish.DishCategoryID);
-            return View(dish);
+            return RedirectToAction("Login", "Account");            
         }
 
         // POST: Dishes/Edit/5
@@ -82,29 +105,37 @@ namespace Thesis.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "Name,Price,DishCategoryID")] Dish dish)
         {
-            if (ModelState.IsValid)
+            if (AuthenticationTools.UserHasRole(currentUserName, "StaffMember,Boss,Admin"))
             {
-                db.Entry(dish).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Entry(dish).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                ViewBag.DishCategoryID = new SelectList(db.DishCategories, "ID", "Name", dish.DishCategoryID);
+                return View(dish);
             }
-            ViewBag.DishCategoryID = new SelectList(db.DishCategories, "ID", "Name", dish.DishCategoryID);
-            return View(dish);
+            return RedirectToAction("Login", "Account");            
         }
 
         // GET: Dishes/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (AuthenticationTools.UserHasRole(currentUserName, "StaffMember,Boss,Admin"))
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Dish dish = db.Dishes.Find(id);
+                if (dish == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(dish);
             }
-            Dish dish = db.Dishes.Find(id);
-            if (dish == null)
-            {
-                return HttpNotFound();
-            }
-            return View(dish);
+            return RedirectToAction("Login", "Account");            
         }
 
         // POST: Dishes/Delete/5
@@ -112,10 +143,14 @@ namespace Thesis.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Dish dish = db.Dishes.Find(id);
-            db.Dishes.Remove(dish);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (AuthenticationTools.UserHasRole(currentUserName, "StaffMember,Boss,Admin"))
+            {
+                Dish dish = db.Dishes.Find(id);
+                db.Dishes.Remove(dish);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            return RedirectToAction("Login", "Account");
         }
 
         protected override void Dispose(bool disposing)
